Story package restructuring
