+++ packages/core/stories/link.doc.stories.mdx
@@ -1,14 +1,12 @@
 import { Meta, ArgsTable, Canvas, Story } from "@storybook/addon-docs";
 import { Link } from "@salt-ds/core";
+import { StackoverflowIcon } from "@salt-ds/icons";
+import { withFlexGap } from "docs/decorators/withFlexGap";
+import HelpAndSupport from "docs/blocks/help-and-support.doc.mdx";
 import {
   CharacteristicUsage,
   CSSClassTable,
 } from "css-variable-docgen-components";
-
-import { withFlexGap } from "docs/decorators/withFlexGap";
-
-import HelpAndSupport from "docs/blocks/help-and-support.doc.mdx";
-import { StackoverflowIcon } from "@salt-ds/icons";
 
 <Meta
   title="Documentation/Core/Link"
