+++ packages/core/stories/text.doc.stories.mdx
@@ -4,7 +4,6 @@ import {
   CharacteristicUsage,
   CSSClassTable,
 } from "css-variable-docgen-components";
-
 import HelpAndSupport from "docs/blocks/help-and-support.doc.mdx";
 
 <Meta
