+++ packages/lab/src/__tests__/__e2e__/tabs/Tabs.accessibility.cy.tsx
@@ -1,5 +1,5 @@
 import { composeStories } from "@storybook/testing-react";
-import * as tabstripStories from "@stories/tabs/tabstrip.cypress.stories";
+import * as tabstripStories from "@stories/tabstrip.cypress.stories";
 
 const { SimpleTabstrip } = composeStories(tabstripStories);
