+++ packages/lab/src/__tests__/__e2e__/tabs/Tabs.cy.tsx
@@ -1,6 +1,6 @@
 import { composeStories } from "@storybook/testing-react";
-import * as tabstripStories from "@stories/tabs/tabstrip.cypress.stories";
-import * as tabsStories from "@stories/tabs/tabs.stories";
+import * as tabstripStories from "@stories/tabstrip.cypress.stories";
+import * as tabsStories from "@stories/tabs.stories";
 import { Tabs, TabPanel } from "@salt-ds/lab";
 import { version } from "react";
