+++ packages/lab/stories/assets/avatar1.png
