+++ packages/lab/stories/components/CollapsibleToolbarItem.tsx
