+++ packages/lab/stories/components/QueryResultTable.css
