+++ packages/lab/stories/components/QueryResultTable.tsx
