+++ packages/lab/stories/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./AdjustableFlexbox";
+export * from "./CloseTabWarningDialog";
+export * from "./CollapsibleToolbarItem";
+export * from "./QueryResultTable";
