--- packages/lab/stories/sample-components/index.ts
@@ -1,1 +1,0 @@
-export * from "./CollapsibleToolbarItem";
