+++ packages/lab/stories/tabs.doc.stories.mdx
@@ -1,5 +1,5 @@
-import { Meta, Story, Canvas } from "@storybook/addon-docs";
 import { useRef, useState } from "react";
+import { Meta, Story, Canvas } from "@storybook/addon-docs";
 import { Link, SaltProvider, Text } from "@salt-ds/core";
 import { EditableLabel, TabPanel, Tabs, Tab, Tabstrip } from "@salt-ds/lab";
