+++ packages/lab/stories/tabstrip.cypress.stories.tsx
@@ -1,9 +1,7 @@
 import { useState } from "react";
 import { Story } from "@storybook/react";
-
 import { Tabstrip, TabstripProps } from "@salt-ds/lab";
-
-import { AdjustableFlexbox, AdjustableFlexboxProps } from "../story-components";
+import { AdjustableFlexbox, AdjustableFlexboxProps } from "./components";
 
 export default {
   title: "Lab/Tabs/Cypress Test Fixtures",
