+++ packages/lab/stories/toolbar.doc.stories.mdx
@@ -7,7 +7,7 @@ import {
   TearOutIcon,
 } from "@salt-ds/icons";
 import { Toolbar } from "@salt-ds/lab";
-import { CollapsibleToolbarItem } from "./sample-components";
+import { CollapsibleToolbarItem } from "./components";
 
 <Meta title="Documentation/Lab/Toolbar" />
